Use segment Keyformat and Keyformatversions on MediaPlaylist.Encode()
Previously, when a segment has different encryption parameters than the
default playlist, the Encode() function uses the segment's parameters for
all parameters but Keyformat and Keyformatversions.

This change uses the segment's Keyformat and Keyformatversions when the
segments are encoded, allowing these parameters to change per segment and
not requiring a playlist to have a key set.