Increment pls version when either KEYFORMAT or KEYFORMATVERSIONS is set
HLS draft RFC (draft-pantos-http-live-streaming-18) section 4.3.2.4
states that the use of KEYFORMAT REQUIRES playlist version 5 or
greater, and also state the use of KEYFORMATVERSIONS requires playlist
version 5 or greater.

Section 7 then states version 5 is required if "The KEYFORMAT and
KEYFORMATVERSIONS attributes of the EXT-X-KEY tag.",

Currently the version number is incremented only if both options are set.

As some devices set one to its default value and omit the other, the
playlist should increment the version number if either of these options
are set.