Correctly decode a Media manifest with BYTERANGE
decodeLineOfMediaPlaylist currently expects a playlist to have
*either* a EXTINF or a EXT-X-BYTERANGE because it's using an
if EXTINF {} else if EXT-X-BYTERANGE {}, this is incorrect as
EXTINF is required and therefore a playlist must always have
an EXTINF, the EXT-X-BYTERANGE is optional.

This change also fixes a bug when an offset given in EXT-X-BYTERANGE
is not reset in its state, so segments without an offset would
incorrect be given the last offset set.

Also includes unit tests to reproduce the issue.