Expose master and media playlist version setter and getter methods
Currently version is controlled via private version functions and
set via the Set* methods when required. This is the ideal method for
most applications, as the version is managed by the library

This change provides the user the ability to check the version as
well as overwrite it, as during some live streams the version number
may have been incremented via the Set methods, but once those
segments have been removed the version number will change during
playback. This isn't ideal as the player may encounter version changes
during playback to versions it doesn't support. This provides
implementors the ability to announce up front what the version number
will be.