Add reader support for EXT-OATCLS-SCTE35 and EXT-X-CUE-OUT
**Breaking Change**: SCTE struct was renamed to SCTE35

**Breaking Change**: MediaSegment.SCTE was renamed to MediaSegment.SCTE35

**Deprecation**: MediaPlaylist.SetSCTE() has been deprecated, use MediaPlaylist.SetSCTE35() instead

SCTE35 (was SCTE) has 3 new fields, SCTE35.Syntax defines the format of the
SCTE data, SCTE35.CueType defines whether the cue point was start/mind/end,
SCTE35.Elapsed can be used for formats that track the elapsed time.

SCTE35 now supports another SCTE-35 syntax we've nicknamed OATCLS (due
to the presence of this string in the tag), see sample-playlists/media-playlist-with-oatcls-scte35.m3u8
for an example. Both reads and writes are supported.