Make comma optional in #EXTINF when strict mode is false (#85)
**This is a potential breaking change**

Previously we would skip a segment which doesn't have a comma in the

For example, the following playlist would have a single nil segment.

 #EXTM3U
 #EXT-X-VERSION:3
 #EXT-X-TARGETDURATION:10
 c_1267327549363093.ts

The new behaviour will cause an error to be returned when parsing if
we're operating in strict mode. If we're not operating in strict mode
the duration will be parsed successfully as 10.000, with an empty title.

Fixes #84.