Adding Head method to get details of first segment
In a sliding window playlist, we need to increase disconitnuity sequence number if the first segment of the playlist is removed. Since, there is no method or attribute exposed to get the first segment details, we are not able to obtain discontinuity flag from the it. Exposing a method to get the "head" segment of the playlist.